removing price step in publish asset
